Add a more complex timeToMine default registry and hotbar support
Previously, the default timeToMine would use a fixed value of 9.
This default is still used if the requisite plugins are not available,
but if voxel-registry is loaded, then it will be used to lookup the
block hardness value. And if voxel-inventory-hotbar is loaded, the held
item will be used to scale the mining time as well.